pallet_bounties::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

accept_curator

Accept the curator role for a bounty. A deposit will be reserved from curator and refund upon successful payout.

approve_bounty

Approve a bounty proposal. At a later time, the bounty will be funded and become active and the original deposit will be returned.

approve_bounty_with_curator

Approve bountry and propose a curator simultaneously. This call is a shortcut to calling approve_bounty and propose_curator separately.

award_bounty

Award bounty to a beneficiary account. The beneficiary will be able to claim the funds after a delay.

claim_bounty

Claim the payout from an awarded bounty after payout delay.

close_bounty

Cancel a proposed or active bounty. All the funds will be sent to treasury and the curator deposit will be unreserved if possible.

extend_bounty_expiry

Extend the expiry time of an active bounty.

propose_bounty

Propose a new bounty.

propose_curator

Propose a curator to a funded bounty.

unassign_curator

Unassign curator from a bounty.