pallet_bounties::pallet::dispatchables

Function unassign_curator

pub fn unassign_curator<T: Config<I>, I: 'static>(bounty_id: BountyIndex)

Unassign curator from a bounty.

This function can only be called by the RejectOrigin a signed origin.

If this function is called by the RejectOrigin, we assume that the curator is malicious or inactive. As a result, we will slash the curator when possible.

If the origin is the curator, we take this as a sign they are unable to do their job and they willingly give up. We could slash them, but for now we allow them to recover their deposit and exit without issue. (We may want to change this if it is abused.)

Finally, the origin can be anyone if and only if the curator is “inactive”. This allows anyone in the community to call out that a curator is not doing their due diligence, and we should pick a new curator. In this case the curator should also be slashed.

Complexity

O(1).

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::unassign_curator.