Trait pallet_bounties::pallet::Config

pub trait Config<I: 'static = ()>: Config + Config<I> {
    type BountyDepositBase: Get<BalanceOf<Self, I>>;
    type BountyDepositPayoutDelay: Get<Self::BlockNumber>;
    type BountyUpdatePeriod: Get<Self::BlockNumber>;
    type CuratorDepositMultiplier: Get<Permill>;
    type CuratorDepositMax: Get<Option<BalanceOf<Self, I>>>;
    type CuratorDepositMin: Get<Option<BalanceOf<Self, I>>>;
    type BountyValueMinimum: Get<BalanceOf<Self, I>>;
    type DataDepositPerByte: Get<BalanceOf<Self, I>>;
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type MaximumReasonLength: Get<u32>;
    type WeightInfo: WeightInfo;
    type ChildBountyManager: ChildBountyManager<BalanceOf<Self, I>>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type BountyDepositBase: Get<BalanceOf<Self, I>>

The amount held on deposit for placing a bounty proposal.

type BountyDepositPayoutDelay: Get<Self::BlockNumber>

The delay period for which a bounty beneficiary need to wait before claim the payout.

type BountyUpdatePeriod: Get<Self::BlockNumber>

Bounty duration in blocks.

type CuratorDepositMultiplier: Get<Permill>

The curator deposit is calculated as a percentage of the curator fee.

This deposit has optional upper and lower bounds with CuratorDepositMax and CuratorDepositMin.

type CuratorDepositMax: Get<Option<BalanceOf<Self, I>>>

Maximum amount of funds that should be placed in a deposit for making a proposal.

type CuratorDepositMin: Get<Option<BalanceOf<Self, I>>>

Minimum amount of funds that should be placed in a deposit for making a proposal.

type BountyValueMinimum: Get<BalanceOf<Self, I>>

Minimum value for a bounty.

type DataDepositPerByte: Get<BalanceOf<Self, I>>

The amount held on deposit per byte within the tip report reason or bounty description.

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The overarching event type.

type MaximumReasonLength: Get<u32>

Maximum acceptable reason length.

Benchmarks depend on this value, be sure to update weights file when changing this value

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type ChildBountyManager: ChildBountyManager<BalanceOf<Self, I>>

The child bounty manager.